+++ src/components/sales-regions/sales-regions.js
@@ -1,7 +1,7 @@
 import * as bootstrap from "bootstrap";
 import "../../style.scss";
 import "./sales-regions.scss";
-import { loadHeader, loadFooter } from "../global/global.js";
+import { loadHeader, loadFooter, loadMap } from "../global/global.js";
 
 function activeLinkNav() {
   const link = document.querySelector(".sales-regions-link ");
@@ -11,5 +11,6 @@ function activeLinkNav() {
 document.addEventListener("DOMContentLoaded", async () => {
   await loadHeader();
   await loadFooter();
+  await loadMap();
   activeLinkNav();
 });
