+++ src/pages/sales-regions/sales-regions.html
@@ -14,7 +14,33 @@
       </div>
     </div>
     <main class="main">
-      <div class="sale-sea-containers"></div>
+      <div class="sale-sea-containers container-fluid">
+        <div class="container container-sale-sea-containers">
+          <div class="row">
+            <h1 class="title-page-text p-0">
+              Продажа морских контейнеров в регионах
+            </h1>
+          </div>
+          <div class="row">
+            <ul class="list-title">
+              <li>10 футов</li>
+              <li>20 футов</li>
+              <li>40 футов</li>
+              <li>45 футов</li>
+              <li>Рефрижераторы</li>
+            </ul>
+          </div>
+          <div class="row">
+            <form
+              class="container-fluid justify-content-start col-3 d-flex justify-content-start mx-0"
+            >
+              <button class="btn btn-primary btn-order" type="button">
+                Заказать
+              </button>
+            </form>
+          </div>
+        </div>
+      </div>
       <div class="clients">
         <div class="container-fluid">
           <div
